Added provisioning checks if packages already installed. Added sass to installation as well.

+++ bootstrap.sh
@@ -6,6 +6,11 @@ dbdirectory=db
 
 exampleconf=/vagrant/"$scriptsdirectory"/apacheconf/example.conf
 
+config_count=0
+
+#############################
+# Create directories needed #
+#############################
 
 # Create the forlder for the websites
 if [ ! -d /vagrant/"$webdirectory" ]; then
@@ -37,41 +42,53 @@ if [ -d /vagrant/"$scriptsdirectory"/apacheconf ]; then
 		if [ "$filename" != "$exampleconf" ]; then
 
 			echo "Copying $filename to /etc/apache2/sites-available/"
-    		sudo cp "$filename" /etc/apache2/sites-available/
+    	sudo cp "$filename" /etc/apache2/sites-available/
     		
-    		filename_without_path=$(basename $filename)
-    		filename_without_ext="${filename_without_path%.*}"
+    	filename_without_path=$(basename $filename)
+    	filename_without_ext="${filename_without_path%.*}"
 
-    		echo "Enabling the site on apache server"
-    		a2ensite "$filename_without_ext"
+    	echo "Enabling the site on apache server"
+    	a2ensite "$filename_without_ext"
 
-    	fi
+    	config_count=$[config_count + 1]
+
+    fi
 
 	done
 fi
 
-echo "Reloading apache configuration"
-sudo service apache2 reload
+
+###############################################################
+# Reload the apache config if there are new virtualhosts      #
+###############################################################
+if [ "$config_count" -gt "0" ]; then
+	echo "Reloading apache configuration"
+	sudo service apache2 reload
+fi
+
 
 #####################
 # Install WP-CLI    #
 #####################
-echo "Fetching WP-CLI"
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-echo "Installign WP-CLI"
-chmod +x wp-cli.phar
-sudo mv wp-cli.phar /usr/local/bin/wp
-echo "WP-CLI installation done. Use wp on commandline"
-wp --info
+if [ ! -f /usr/local/bin/wp ]; then
+	echo "Fetching WP-CLI"
+	curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+	echo "Installign WP-CLI"
+	chmod +x wp-cli.phar
+	sudo mv wp-cli.phar /usr/local/bin/wp
+	echo "WP-CLI installation done. Use wp on commandline"
+	wp --info
+fi
 
 
 #######################
 # Install Composer    #
 #######################
-echo "Installing composer"
-curl -sS https://getcomposer.org/installer | php
-sudo mv composer.phar /usr/local/bin/composer
-echo "Composer installed"
+if [ ! -f /usr/local/bin/composer ]; then
+	echo "Installing composer"
+	curl -sS https://getcomposer.org/installer | php
+	sudo mv composer.phar /usr/local/bin/composer
+fi
 
 
 #######################
@@ -81,23 +98,28 @@ echo "Composer installed"
 # apt version does    #
 # not work            #
 #######################
-echo "Installing nodejs and npm"
-curl -sL https://deb.nodesource.com/setup | sudo bash -
-sudo apt-get install -y nodejs
+if [ ! -f /usr/bin/npm ]; then
+	echo "Installing nodejs and npm"
+	curl -sL https://deb.nodesource.com/setup | sudo bash -
+	sudo apt-get install -y nodejs
+fi
 
 #######################
 # Install grunt       #
 #######################
-echo "Installing grunt"
-sudo npm install -g grunt-cli
+if [ ! -f /usr/bin/grunt ]; then
+	echo "Installing grunt"
+	sudo npm install -g grunt-cli
+fi
 
 
 #######################
 # Install SASS        #
 #######################
-sudo gem install sass
-
-
+if [ ! -f /opt/vagrant_ruby/bin/sass ]; then
+	echo "Installing sass"
+	gem install sass
+fi
 
 
 #######################
